added health tracking system visible in the dev console, no styling representation yet.

+++ public/main.js
@@ -1,7 +1,9 @@
 const numberOfDrones = 15;
+let playerHealth = 100;
 let invasionForce = [];
 
 const gameStart = () => {
+  playerHealth = 10;
   invasionForce = [];
   spawnInvasionForce();
   newRound();
@@ -16,9 +18,31 @@ const spawnInvasionForce = () => {
 };
 
 const newRound = () => {
+  playerHealth <= 0 ? loseGame() : invasionForce.length == 0 ? winGame() : null;
   document.getElementById("leader").innerHTML = "";
   document.getElementById("swarm").innerHTML = "";
   showInvasionForce();
+};
+
+const loseGame = () => {
+  console.log("You Lose!");
+};
+
+const winGame = () => {
+  console.log("You Win!");
+};
+
+const updateHealthBar = attackType => {
+  if (invasionForce.length > 0) {
+    attackType === "single-shot"
+      ? (playerHealth -= 5)
+      : attackType === "scatter-shot"
+      ? (playerHealth -= 10)
+      : attackType === "bomb"
+      ? (playerHealth -= 15)
+      : (playerHealth -= 30);
+    console.log(playerHealth);
+  }
 };
 
 const showInvasionForce = () => {
@@ -31,48 +55,60 @@ const showInvasionForce = () => {
   });
 };
 
-const shoot = () => {
-  let randomIndex = Math.floor(Math.random() * invasionForce.length);
-  invasionForce[randomIndex].hp -= 10;
-  checkMotherShip();
-  deleteDead();
-  newRound();
+const singleShot = () => {
+  if (invasionForce.length > 0 && playerHealth > 0) {
+    let randomIndex = Math.floor(Math.random() * invasionForce.length);
+    invasionForce[randomIndex].hp -= 10;
+    checkMotherShip();
+    deleteDead();
+    updateHealthBar("single-shot");
+    newRound();
+  }
 };
 
 const scatterShot = () => {
-  let randomIndex1 = Math.floor(Math.random() * invasionForce.length);
-  let randomIndex2 = Math.floor(Math.random() * invasionForce.length);
-  let randomIndex3 = Math.floor(Math.random() * invasionForce.length);
-  let randomIndex4 = Math.floor(Math.random() * invasionForce.length);
+  if (invasionForce.length > 0 && playerHealth > 0) {
+    let randomIndex1 = Math.floor(Math.random() * invasionForce.length);
+    let randomIndex2 = Math.floor(Math.random() * invasionForce.length);
+    let randomIndex3 = Math.floor(Math.random() * invasionForce.length);
+    let randomIndex4 = Math.floor(Math.random() * invasionForce.length);
 
-  invasionForce[randomIndex1].hp -= 10;
-  invasionForce[randomIndex2].hp -= 10;
-  invasionForce[randomIndex3].hp -= 5;
-  invasionForce[randomIndex4].hp -= 5;
-  checkMotherShip();
-  deleteDead();
-  newRound();
+    invasionForce[randomIndex1].hp -= 10;
+    invasionForce[randomIndex2].hp -= 10;
+    invasionForce[randomIndex3].hp -= 5;
+    invasionForce[randomIndex4].hp -= 5;
+    checkMotherShip();
+    deleteDead();
+    updateHealthBar("scatter-shot");
+    newRound();
+  }
 };
 
 const bomb = () => {
-  let randomIndex = Math.floor(Math.random() * invasionForce.length);
-  invasionForce[randomIndex + 1].hp -= 10;
-  invasionForce[randomIndex + 2].hp -= 10;
-  invasionForce[randomIndex].hp -= 10;
-  invasionForce[randomIndex - 1].hp -= 10;
-  invasionForce[randomIndex - 2].hp -= 10;
-  checkMotherShip();
-  deleteDead();
-  newRound();
+  if (invasionForce.length > 0 && playerHealth > 0) {
+    let randomIndex = Math.floor(Math.random() * invasionForce.length);
+    invasionForce[randomIndex + 1].hp -= 10;
+    invasionForce[randomIndex + 2].hp -= 10;
+    invasionForce[randomIndex].hp -= 10;
+    invasionForce[randomIndex - 1].hp -= 10;
+    invasionForce[randomIndex - 2].hp -= 10;
+    checkMotherShip();
+    deleteDead();
+    updateHealthBar("bomb");
+    newRound();
+  }
 };
 
 const nuke = () => {
-  invasionForce.map(alien => {
-    alien.hp -= 15;
-  });
-  checkMotherShip();
-  deleteDead();
-  newRound();
+  if (invasionForce.length > 0 && playerHealth > 0) {
+    invasionForce.map(alien => {
+      alien.hp -= 15;
+    });
+    checkMotherShip("nuke");
+    updateHealthBar();
+    deleteDead();
+    newRound();
+  }
 };
 
 const checkMotherShip = () => {
